Refactor integration tests setup and teardown functions for improved organization
